// a3front/src/modules/negotiations/supplier-new/enums/form-component.enum.ts
export enum FormComponentEnum {
  CLASSIFICATION_SUPPLIER = 'classificationSupplier',
  GENERAL_INFORMATION = 'generalInformation',
  COMMERCIAL_LOCATION = 'commercialLocation',
  COMMERCIAL_INFORMATION = 'commercialInformation',
  PLACE_OPERATION = 'placeOperation',
  CONTACT_INFORMATION = 'contactInformation',
  BASIC_INFORMATION = 'basicInformation',
  SUNAT_INFORMATION = 'sunatInformation',

  MODULE_LIST_UNITS = 'listUnits',
  MODULE_SERVICES = 'services',
  MODULE_POLICIES = 'policies',
  MODULE_CONTACT = 'contact',
  MODULE_DATA_SHEET = 'dataSheet',
  MODULE_SUNAT_INFORMATION = 'moduleSunatInformation',
  MODULE_TREASURY_BANK_INFORMATION = 'moduleTreasuryBankInformation',
  MODULE_TREASURY_BENEFICIARIES = 'moduleTreasuryBeneficiaries',
  MODULE_TREASURY_ACCOUNT_CONTABLE = 'moduleTreasuryAccountContable',
  MODULE_ACCOUNTING_SUNAT_INFORMATION = 'moduleAccountingSunatInformation',

  COMMERCIAL_ATTRACTIONS = 'touristAttractions',
  COMMERCIAL_LANGUAGE = 'languageInformation',
}